Added namespace to Extensions test
