+++ tests/Assets/Extensions/ExtensionMock2.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace LaravelDoctrineTest\ORM\Assets\Extensions;
+
+use Doctrine\Common\EventManager;
+use Doctrine\ORM\EntityManagerInterface;
+use LaravelDoctrine\ORM\Extensions\Extension;
+
+class ExtensionMock2 implements Extension
+{
+    public function addSubscribers(EventManager $manager, EntityManagerInterface $em): void
+    {
+    }
+
+    /**
+     * @return mixed[]
+     */
+    public function getFilters(): array
+    {
+        return [];
+    }
+}
